feat(ui): single floating widget (Chat/Queue)
- 1 seule bulle + header switch Chat/Queue\n- QueueMonitor réutilisé comme panel (sans wrapper/header)\n- TODO: notes sur la responsiveness du widget

docs: update BRANCH status for 2E + widget

docs: clarify widget desktop responsive option
