+++ ui/src/lib/components/ChatPanel.svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+  // Placeholder minimal: l'UI chat (sessions/messages/composer) arrive ensuite.
+</script>
+
+<div class="p-4 text-sm text-slate-600">
+  <div class="font-medium text-slate-900 mb-1">Chat</div>
+  <p class="text-xs text-slate-500">
+    UI en cours de construction. Le streaming passera via <code>/api/v1/streams/sse</code> et <code>streamId = assistantMessageId</code>.
+  </p>
+</div>
+
+
