+++ ui/src/routes/+layout.svelte
@@ -7,7 +7,7 @@
   import Header from '$lib/components/Header.svelte';
   import Toast from '$lib/components/Toast.svelte';
   import NavigationGuard from '$lib/components/NavigationGuard.svelte';
-  import QueueMonitor from '$lib/components/QueueMonitor.svelte';
+  import OperationsWidget from '$lib/components/OperationsWidget.svelte';
   import '$lib/i18n';
   import { initializeSession, session } from '$lib/stores/session';
 
@@ -174,5 +174,5 @@
   </main>
   <Toast />
   <NavigationGuard />
-  <QueueMonitor />
+  <OperationsWidget />
 </div>
